requirements.txt
rembg
Pillow>=9.0.0,<11.0.0
Flask==2.3.2
flask-cors==3.0.10